Export metrics with timestamps (#16)
By switching to the custom metric exporter, we can add timestamps to the
metric streams. This allows us to run multiple celerymon instances since
metric streams can be aggregated.

For example, if we have two celerymon instances:

* instance1: active_task_count{task_name='task1'} = 3
* instance2: active_task_count{task_name='task1'} = 5

Because the value update timings are not aligned, the there can be a
discrepancy between these two instances. This gets more obvious when one
of the instances is not updating the data for a while due to an issue.
By having a timestamp attached to the metric streams, the collector can
tell what is the latest value for the given metric.